refactor: move hooks to reducers folder
